+++ common/src/main/java/ru/dimaskama/webcam/net/S2CPacket.java
@@ -2,6 +2,7 @@ package ru.dimaskama.webcam.net;
 
 import ru.dimaskama.webcam.net.packet.Packet;
 
+import java.nio.ByteBuffer;
 import java.util.UUID;
 
 // Utility class
@@ -9,10 +10,14 @@ public final class S2CPacket {
 
     public static byte MAGIC = (byte) 0b11101110;
 
-    public static byte[] create(byte[] encrypted) throws Exception {
-        byte[] result = new byte[1 + encrypted.length];
-        result[0] = MAGIC;
-        System.arraycopy(encrypted, 0, result, 1, encrypted.length);
+    public static byte[] create(UUID playerUuid, byte[] encrypted) throws Exception {
+        byte[] result = new byte[17 + encrypted.length];
+        ByteBuffer buffer = ByteBuffer.wrap(result);
+        buffer
+                .put(MAGIC)
+                .putLong(playerUuid.getMostSignificantBits())
+                .putLong(playerUuid.getLeastSignificantBits())
+                .put(encrypted);
         return result;
     }
 
@@ -20,7 +25,8 @@ public final class S2CPacket {
         if (data[0] != MAGIC) {
             throw new IllegalArgumentException("Invalid packet");
         }
-        return Packet.decrypt(data, 1, data.length - 1, secret);
+        // Skip UUID
+        return Packet.decrypt(data, 17, data.length - 17, secret);
     }
 
 }