+++ velocity/src/template/java/ru/dimaskama/webcam/velocity/WebcamVelocityVersion.java
@@ -1,0 +1,7 @@
+package ru.dimaskama.webcam.velocity;
+
+public class WebcamVelocityVersion {
+
+    public static final String VERSION = "${mod_version}";
+
+}